requirements.txt
Flask~=3.0.2
chromadb~=0.4.24
pillow~=10.2.0
joblib~=1.3.2